Renamed new_interface to interface
